Update the contributing guide
Here is a summary of a conversation I had with @Inwerpsel, where he helped me setting up my development environment. We agreed that this information is useful to have in the docs. Though, I'm not really sure where to put this. I put information in the contribute page, but please let me know if there is somewhere better to put it.